+++ src/Updater.Publish/AssemblyInfo.fs
@@ -5,11 +5,11 @@ open System.Reflection
 [<assembly: AssemblyProductAttribute("Updater")>]
 [<assembly: AssemblyCompanyAttribute("Dzmitry Martynau")>]
 [<assembly: AssemblyDescriptionAttribute("Deploy, update and publish tools for Windows client applications.")>]
-[<assembly: AssemblyVersionAttribute("0.2.10")>]
-[<assembly: AssemblyFileVersionAttribute("0.2.10")>]
-[<assembly: AssemblyInformationalVersionAttribute("0.2.10")>]
+[<assembly: AssemblyVersionAttribute("0.2.11")>]
+[<assembly: AssemblyFileVersionAttribute("0.2.11")>]
+[<assembly: AssemblyInformationalVersionAttribute("0.2.11")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "0.2.10"
-    let [<Literal>] InformationalVersion = "0.2.10"
+    let [<Literal>] Version = "0.2.11"
+    let [<Literal>] InformationalVersion = "0.2.11"